Failing test: video with cumulative views
Facebook is gradually changing the HTML of the video pages so that
views are displayed cumulatively in a pop-up. For example:

https://www.facebook.com/video.php?v=203203106739575

Unfortunately this breaks Funky.
And unfortunately this is not for **every** video (yet?) so we have
to be able to parse both the new and old HTML.

This PR simply adds a test to show the error in Travis.