`Funky::Video.where(id: id)` should never error.
Before, if a single video id was passed into Funky::Video.where,
and that video could not be accessed, it would raise ContentNotFound.
That goes against the readme and is not consistent with when we pass
multiple ids to the where clause. The where clause should simply
ignore badd video ids. If it's a single video id, it should return
and empty array.