Change how to scrape a page of a video
Facebook goes to a 'not found' page when you try to visit a video
with invalid video id on URL. So for scraping reason, we need to
distinguish this 'not found' page from actual video pages.

The previous method was also filtering some actual videos (because
Facebook changed their HTML format) so, this is a try to filter
only 'not found' pages.